Working model. First stab

+++ light_malib/envs/LMAPF/_env.py
@@ -224,11 +224,11 @@ class LMAPFEnv(BaseEnv):
             heuristics[~masks]=-1
                 
         if guiding_actions is not None:
-            if torch.any(heuristics[torch.arange(len(guiding_actions),dtype=torch.int32,device=self.device),guiding_actions]<0):
+            if torch.any(heuristics[torch.arange(len(guiding_actions),dtype=torch.long,device=self.device),guiding_actions.long()]<0):
                 Logger.error("err! this may sometimes happen if the probability is not safely clipped above 1e-9, for example.")
             # TODO: set to 0 has the potential bug when agent will arrive the goal?
-            heuristics[torch.arange(len(guiding_actions),dtype=torch.int32,device=self.device),guiding_actions]=torch.minimum(
-                heuristics[torch.arange(len(guiding_actions),dtype=torch.int32,device=self.device),guiding_actions],
+            heuristics[torch.arange(len(guiding_actions),dtype=torch.long,device=self.device),guiding_actions.long()]=torch.minimum(
+                heuristics[torch.arange(len(guiding_actions),dtype=torch.long,device=self.device),guiding_actions.long()],
                 torch.tensor(0,device=self.device,dtype=torch.float32)
             )
             
@@ -245,7 +245,7 @@ class LMAPFEnv(BaseEnv):
         return self._initialized
     
     def disable_agents(self):
-        mask=self.corner_graph[self._target_positions[...,0],self._target_positions[...,1]]
+        mask=self.corner_graph[self._target_positions[...,0].long(),self._target_positions[...,1].long()]
         disabled=torch.zeros_like(self.priorities,dtype=torch.bool,device=self.device)
         disabled[mask]=True
         self.target_positions=self._target_positions.clone()
@@ -399,11 +399,11 @@ class LMAPFEnv(BaseEnv):
 
 
         self.actions=actions
-        
-        self.action_cnts[torch.arange(self.num_robots,device=self.device),actions]+=1
+
+        self.action_cnts[torch.arange(self.num_robots,dtype=torch.long,device=self.device),actions.long()]+=1
         self.episode_log.add_actions(actions)
         
-        movements=self.movements[actions]
+        movements=self.movements[actions.long()]
         next_positions=self.curr_positions+movements
         global_timer.time("pibt_s","pibt_e","pibt")
         
@@ -620,13 +620,13 @@ class LMAPFEnv(BaseEnv):
         # Feature 1: other static obstacles
         # num_robots, FOV_height, FOV_width
         # pad with 1, meaining obstacle
-        local_static_obstacle_map=self.padded_graph[offsetted_local_views[...,0],offsetted_local_views[...,1]]
+        local_static_obstacle_map=self.padded_graph[offsetted_local_views[...,0].long(),offsetted_local_views[...,1].long()]
         
         # Feature 2: other agents
         global_agents_map=torch.zeros_like(self.padded_graph,dtype=torch.float32)
-        global_agents_map[self.curr_positions[:,0]+self.FOV_height//2,self.curr_positions[:,1]+self.FOV_width//2]=self.priorities
+        global_agents_map[(self.curr_positions[:,0]+self.FOV_height//2).long(),(self.curr_positions[:,1]+self.FOV_width//2).long()]=self.priorities
         # num_robots, FOV_height, FOV_width
-        local_agents_map=global_agents_map[offsetted_local_views[...,0],offsetted_local_views[...,1]]
+        local_agents_map=global_agents_map[offsetted_local_views[...,0].long(),offsetted_local_views[...,1].long()]
         # set its own position to 0
         # local_agents_map[:,self.FOV_height//2,self.FOV_width//2]=0
         # NOTE(rivers): >0 is a bug before. we fix it.
@@ -671,9 +671,9 @@ class LMAPFEnv(BaseEnv):
         global_observations=torch.zeros(size=(1,self.num_global_channels,self.map.height,self.map.width),dtype=torch.float32,device=self.device)
         
         global_observations[0,0]=self.graph.float()
-        global_observations[0,1,self.curr_positions[:,0],self.curr_positions[:,1]]=1
-        global_observations[0,2,self.target_positions[:,0],self.target_positions[:,1]]=1        
-        
+        global_observations[0,1,self.curr_positions[:,0].long(),self.curr_positions[:,1].long()]=1
+        global_observations[0,2,self.target_positions[:,0].long(),self.target_positions[:,1].long()]=1
+
         return observations, global_observations
     
     def get_action_masks(self):
@@ -686,7 +686,7 @@ class LMAPFEnv(BaseEnv):
         
         # check if out of bound and if collide with static obstacles
         # 0-valid, 1-invalid
-        action_masks=self.padded_graph[next_positions[...,0],next_positions[...,1]]
+        action_masks=self.padded_graph[next_positions[...,0].long(),next_positions[...,1].long()]
         
         # 1-valid, 0-invalid
         action_masks=1-action_masks.float()
@@ -719,9 +719,9 @@ class LMAPFEnv(BaseEnv):
         
         offseted_curr_positions=self.curr_positions+self.padded_graph_offsets
         
-        reward_map[offseted_curr_positions[:,0],offseted_curr_positions[:,1]]=rewards
-        agent_map[offseted_curr_positions[:,0],offseted_curr_positions[:,1]]=1
-        
+        reward_map[offseted_curr_positions[:,0].long(),offseted_curr_positions[:,1].long()]=rewards
+        agent_map[offseted_curr_positions[:,0].long(),offseted_curr_positions[:,1].long()]=1
+
         kernel=torch.ones((3,3),dtype=torch.float32,device=self.device)
         
         assert kernel.shape[0]==kernel.shape[1] and kernel.shape[0]%2==1 and kernel.shape[1]%2==1
@@ -733,10 +733,10 @@ class LMAPFEnv(BaseEnv):
         
         team_rewards = team_rewards.reshape(*reward_map.shape)
         num_agents = num_agents.reshape(*agent_map.shape)
-        
-        team_rewards = team_rewards[offseted_curr_positions[:,0],offseted_curr_positions[:,1]]
-        num_agents = num_agents[offseted_curr_positions[:,0],offseted_curr_positions[:,1]]
-        
+
+        team_rewards = team_rewards[offseted_curr_positions[:,0].long(),offseted_curr_positions[:,1].long()]
+        num_agents = num_agents[offseted_curr_positions[:,0].long(),offseted_curr_positions[:,1].long()]
+
         # average moving speed
         team_rewards = (team_rewards) / (num_agents+1e-9)
